requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-multipart
